regular_parser/requirements.txt
vk_api
requests
tqdm
psycopg2-binary
python-dotenv
transformers
Pillow
torch
aiohttp

